SMP-501 and SMP-503 Modifications done for Partner search using Date range and Biller search using Date range

+++ Commons/UICore/src/main/java/com/mfino/uicore/fix/processor/impl/BranchCodeProcessorImpl.java
@@ -89,7 +89,7 @@ public class BranchCodeProcessorImpl extends BaseFixProcessor implements BranchC
 	
 	private void updateEntity(Role role, CMJSRole.CGEntries e) {		
 		if(e.getPriorityLevel() != null) {
-			role.setPrioritylevel(e.getPriorityLevel().shortValue());
+			role.setPrioritylevel(e.getPriorityLevel());
 		}
 		if(e.getDisplayText() != null) {
 			role.setDisplaytext(e.getDisplayText());
